Инты сделал даблами, сделал проверку результата на дабловость, убрал цикл из метода проверки ответа

+++ src/com/startjava/lesson_2_3_4/calculator/Calculator.java
@@ -22,7 +22,7 @@ public class Calculator {
                 result = a * b;
                 break;
             case '/':
-                result = (double) a / b;
+                result = a / b;
                 break;
             case '^':
                 result = Math.pow(a, b);
